Move over tests and arbitrary instances from the original testsuite
